+++ FizzimGui.java
@@ -233,7 +233,7 @@ public class FizzimGui extends javax.swing.JFrame {
 		globalList.get(4).add(new ObjAttribute("equation", "1", 1,
 				"def_type","",Color.black,"","",editable));
 
-		globalList.get(4).add(new ObjAttribute("priority", "1000", 1,
+		globalList.get(4).add(new ObjAttribute("priority", "1000", ObjAttribute.NONDEFAULT,
 				"","",Color.black,"","",editable));
 	}